Move serializers to own package
